ui_test/use_playwright/requirements.txt
pytest-playwright
pytest
faker
pytest-xdist
allure-pytest